+++ src/MongoDB.Driver/AggregateFluentBase.cs
@@ -142,7 +142,7 @@ namespace MongoDB.Driver
         public abstract IAggregateFluent<TNewResult> Group<TNewResult>(ProjectionDefinition<TResult, TNewResult> group);
 
         /// <inheritdoc />
-        public abstract IAggregateFluent<TResult> Limit(int limit);
+        public abstract IAggregateFluent<TResult> Limit(long limit);
 
         /// <inheritdoc />
         public virtual IAggregateFluent<TNewResult> Lookup<TForeignDocument, TNewResult>(string foreignCollectionName, FieldDefinition<TResult> localField, FieldDefinition<TForeignDocument> foreignField, FieldDefinition<TNewResult> @as, AggregateLookupOptions<TForeignDocument, TNewResult> options)
@@ -261,7 +261,7 @@ namespace MongoDB.Driver
         }
 
         /// <inheritdoc />
-        public abstract IAggregateFluent<TResult> Skip(int skip);
+        public abstract IAggregateFluent<TResult> Skip(long skip);
 
         /// <inheritdoc />
         public abstract IAggregateFluent<TResult> Sort(SortDefinition<TResult> sort);
